requirements.txt
agno==1.7.11
duckduckgo_search==8.1.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.30.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
python-dotenv==1.1.1

